% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Barriers.R
\name{AddBeringStrait}
\alias{AddBeringStrait}
\title{Add Bering Strait}
\usage{
AddBeringStrait(P, firstStep)
}
\arguments{
\item{P}{A list of parameters.}

\item{firstStep}{The local directions created by OneStepDirections().}
}
\description{
Removes connections at the FirstStep stage of the Local structure to create "barriers" between cells.  Bering Strait Barriers are designed to create structures similar to the Bering Strait entering North America, going through Central America, then opening up into South America.
}
\keyword{Barriers}
