% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{AddSegment}
\alias{AddSegment}
\title{Add Segment}
\usage{
AddSegment(P, a, b, top = FALSE)
}
\arguments{
\item{P}{A list of parameters.}

\item{Data}{The Pre or Post output from an HBC simulation.}
}
\description{
Creates lines to show the Bering Straight borders.
}
\keyword{Plotting}
