% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{BeringStraightPlot}
\alias{BeringStraightPlot}
\title{Bering Straight Plot}
\usage{
BeringStraightPlot(P, Data)
}
\arguments{
\item{P}{A list of parameters.}

\item{Data}{The Pre or Post output from an HBC simulation.}
}
\description{
Creates a plot that shows the Bring strait boundaries.
}
\keyword{Plotting}
