% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Directions.R
\name{CardinalDirections}
\alias{CardinalDirections}
\title{Cardinal Directions}
\usage{
CardinalDirections(target, R, start, round, South, North, East, West, SE,
  NE, SW, NW)
}
\arguments{
\item{target}{The territory around which to get local territoies.}

\item{R}{The number of rows.}

\item{start}{How much to offset numbers (for phoneme structures).}

\item{round}{Whether to get a "round" set of territories (N, S, E, W only) for phonemes or a square set of territories (includes diagonals) for distance.}

\item{SE}{Whether to get the southeasrern territory.}

\item{NE}{Whether to get the northestern territory.}

\item{SW}{Whether to get the southwestern territory.}

\item{NW}{Whether to get the northwestern territory.}

\item{south}{Whether to get the southern territory.}

\item{north}{Whether to get the northern territory.}

\item{east}{Whether to get the eastern territory.}

\item{west}{Whether to get the western territory.}
}
\description{
Calculates the terrritory indicies of locations around a target territory (also used for phoneme relatedness in the same way).
}
\keyword{Directions}
