% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_ParametersSimulation.R
\name{DefineParameters}
\alias{DefineParameters}
\title{Define Parameters}
\usage{
DefineParameters(Rows = 40, Cols = 50, ChanceExpand = 0.8,
  PopulationStartIndex = c(1, 2), NumPopulationPhonemes = rep(NA,
  length(PopulationStartIndex)), UsePopSize = TRUE,
  IndividualsStEmSuEM = c(1000, 10, 20, NA), MutationRate = 0.15,
  PhonemeDitribution = c(12, 24, 133), Consonants = 750,
  Vowels = 100, MinConsonant = 6, MinVowel = 6,
  PhonemeProbabilityType = "RealMimic", GrowthMax = 5, GrowthMin = 5,
  Barriers = FALSE, BarrierLength = 30, BarrierBreaks = 4,
  MutationTypeChance = rep(1/5, 5), HorizontalRate = 0.1,
  Bias = TRUE, Steps = 1, HorizontalLocal = TRUE,
  NumberRandomHorizontal = 8, UpRoot = TRUE, Death = TRUE,
  Bering = FALSE, BeringLength = 20, MigrationSimSteps = 300,
  StationaryMutationRate = 0.01, UpRootMutationRate = 0.01,
  HorizontalSimSteps = 400, Waves = FALSE, Seed = NA)
}
\arguments{
\item{Rows}{The number of rows in the world matrix.}

\item{Cols}{The number of columns in the world matrix.}

\item{ChanceExpand}{The chance that a population will either move or send off a group of individuals to found a new population.}

\item{PopulationStartIndex}{The position in the matrix where each seed population starts.  The number of seed populations is defined by the number of starting indicies.}

\item{NumPopulationPhonemes}{The number of phonemes in each starting population.  If set to NA, this is decided by sampling from a distribution with min, mode, and made on the values from the PhonemeDistribution arguement.}

\item{UsePopSize}{Whether to take into account the the population size (number of people) when making decisions about moving, immegrating, the mutation rate ,and phoneme loss/addition biases.}

\item{IndividualsStEmSuEM}{Four related parameters: 1) The number of individuals a seed population starts with, 2) the minumum number of individuals required to make a founder party to settle a new territory, 3) the minumum number of individuals that must stay behind when a founder party is sent off, and 4) the maximum number of individuals allowed to be in one founder party.}

\item{MutationRate}{The rate at which phonemes mutate.  E.g., if MutationRate==0.1, each phoneme in a populatiosn phoneme inventory has a 10\% chance to mutate.  Note that when usePopSize==TRUE, this is better conceptualized as teh maximum mutation rate (larger populations have lower mutation rates).}

\item{Consonants}{The number of possible consonants in existence.  Default based on real phoneme data.}

\item{Vowels}{The number of possible vowels in existence. Default based on real phoneme data.}

\item{MinConsonant}{The minumum number a consonants that can be in a population's phoneme inventory. Default based on real phoneme data.}

\item{MinVowel}{The minumum number a vowels that can be in a population's phoneme inventory. Default based on real phoneme data.}

\item{PhonemeProbabilityType}{The method by which phoneme probabilities are established.  Can be Real (uses teh real data verbatim, and requires the correct number of cosonants and phonemes), RealMimic (uses teh real data to generate a new distribution of probability similar ot the real data, can be used with any number of phonemes), Equal (all phonemes are equally liekly to be known), Frequency (based on how common phonemes are across populations in teh simulation), or Random (randomly generated).}

\item{GrowthMax}{When an integer, the max number of individuals added or lost from populations every time step.  When a fraction, the percent that population size increases or decerases each timestep.  Can be positive or negative.  Also note that is this an integer, but GrowthMin is not, this will be considered a percentage (i.e. 2 = 200\%).}

\item{GrowthMin}{When an integer, the min number of individuals added or lost from each population every time step.  When a fraction, the percent that a population size increases or decreases each timestep. Can be positive or negative. Also note that is this an integer, but GrowthMax is not, this will be considered a percentage (i.e. 2 = 200\%).}

\item{Barriers}{Whether to create "snake barriers" that limit the direction of migration in the matrix.}

\item{BarrierLength}{The width of snake barriers.}

\item{BarrierBreaks}{The height of the space between snake barriers.}

\item{MutationTypeChance}{The chance that each mutation type occurs.  1) Add, 2) Lose, 3) Split, 4) Join, and 5) Shift.}

\item{HorizontalRate}{The fraction of the population that attempts to modify its phoneme inventory every horizontal timestep.}

\item{Bias}{Whether to randomly bias mutations towards either  gains or losses when populations are small.  Set to true based on previously published data.}

\item{Steps}{The number of distance steps away from a target location that are considered "local."  Includes all 8 cardinal and ordinal directions around a target, so the local area is always a rectangle around the target location.}

\item{HorizontalLocal}{Whether horizontal transfer occurs between local populations or globally.  Set to FALSE as a control, as global horizontal transfer should abolish local patterns.}

\item{NumberRandomHorizontal}{The number of locations to compare when HorizontalLocal==FALSE.  Should be 8 when Steps==1, 24 when steps==2, 48 when Steps=3, ect.}

\item{UpRoot}{Whether established populations can move (TRUE) or they remain in place for the entire simulation (FALSE).}

\item{Death}{Whether populations can die out.}

\item{Bering}{Whether to employ barriers that mimick the Bering Strait and Americas.}

\item{BeringLength}{An integer of length 0 to 24 that degines how long the berring straight is}

\item{MigrationSimSteps}{The number of time steps to run each wave of migration.}

\item{StationaryMutationRate}{The rate at which phonemes evolve in stationary populations during migration.}

\item{UpRootMutationRate}{The rate at which phonemes evolve in populations that move as a unit during migration.}

\item{HorizontalSimSteps}{The number of time steps to spend on horizontal transfer.}

\item{Waves}{Whether migration occurs in waves or all seed populations are added at the same time.  If TRUE, there is one wave for each seed population.}

\item{Seed}{Sets a seed for reproducibility if an integer instead of NA.}

\item{PhonemeDistribution}{The 1) min, 2) mode, and 3) max number of phonemes a population can have when sampling for seed population sizes and when preventing languages from gaining or losing too many phonemes. Defaults based on real phoneme data.}
}
\description{
Creates a parameter data structure for running simulations.
}
\keyword{SimParam}
