% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_GenPopulation.R
\name{Emigrate}
\alias{Emigrate}
\title{Emigrate}
\usage{
Emigrate(P, occupied, local, populations)
}
\arguments{
\item{P}{A list of parameters.}

\item{occupied}{The territories with a population on them.}

\item{local}{The local territories data structure.}

\item{populations}{The data for all exising populations.}
}
\description{
Picks which populations migrate, whether the entire population migrates or a founder party is sent off, and where the population migrates to.  Allows only one population to enter a territory.  When multiple populations attempt to enter the same territory, one is randomly chosen to do so while the rest stay put.
}
\keyword{Population_Dynamics}
