% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Phonemes.R
\name{GenerateSeedLanguage}
\alias{GenerateSeedLanguage}
\title{Generate Seed Language}
\usage{
GenerateSeedLanguage(P, phonemeProbab, seedNum)
}
\arguments{
\item{P}{A list of parameters.}

\item{phonemeProbab}{The probability of gaining each phoneme in the population.}

\item{seedNum}{Which population seed is having it's language generated.}
}
\description{
Creates a language for a seed population.
}
\keyword{Phonemes}
