% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_GenLanguage.R
\name{GetAMutation}
\alias{GetAMutation}
\title{Get A Mutation}
\usage{
GetAMutation(P, phonemes, phonemeProbab, phonemeRelatedness, gain)
}
\arguments{
\item{P}{A list of parameters.}

\item{phonemes}{The phonemes currently in the language.}

\item{phonemeProbab}{The probability of gaining each phoneme in the population.}

\item{phonemeRelatedness}{The phoneme relatedness list.}

\item{gain}{Whether to bias the mutations towards gaining syllables (TRUE) or losing them (FALSE).}
}
\description{
Returns a new Add, Loss, Split, Join, or Shift mutation.  Biases the mutations towards gaining or losing syllables when the parameter Bias==TRUE.
}
\keyword{Language}
