% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{GetColorDistribution}
\alias{GetColorDistribution}
\title{Get Colors}
\usage{
GetColorDistribution(P, Data, i, colors = c("coral1", "coral4"))
}
\arguments{
\item{P}{A list of parameters.}

\item{Data}{The Pre or Post output from an HBC simulation.}

\item{colors}{A vector of beginning and ending colors for a gradient.}
}
\description{
Creates a color gradient.
}
\keyword{Plotting}
