% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{Getcolors}
\alias{Getcolors}
\title{Get Colors}
\usage{
Getcolors(P, Data, i, colors = c("coral1", "coral4"))
}
\arguments{
\item{P}{A list of parameters.}

\item{Data}{The Pre or Post output from an HBC simulation.}
}
\description{
Creates a color gradient.
}
\keyword{Plotting}
