% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{GetDist}
\alias{GetDist}
\title{Get Distance}
\usage{
GetDist(P, point1, point2)
}
\arguments{
\item{P}{A list of parameters.}

\item{point1}{One territory location.}

\item{point2}{Another territory location.}
}
\description{
Get the euclidian distance between two territories.
}
\keyword{Plotting}
