% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Directions.R
\name{GetFactorDim}
\alias{GetFactorDim}
\title{Get Factor Dimentions}
\usage{
GetFactorDim(nPhonemes)
}
\arguments{
\item{nPhonemes}{The number of Phonemes (vowels or consonants).}
}
\description{
Given a number of consomants or vowel, create a datastructure that is as square as possible.
}
\keyword{Directions}
