% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_GenPopulation.R
\name{GetImmigrants}
\alias{GetImmigrants}
\title{Get Immigrants}
\usage{
GetImmigrants(P, occupied, local, populations)
}
\arguments{
\item{P}{A list of parameters.}

\item{occupied}{The territories with a population on them.}

\item{local}{The local territories data structure.}

\item{populations}{The data for all exising populations.}
}
\description{
Tests which populations immigrate, removes those that did not migrate from the migraton data.  Splits the data into populations that sent off founder parties and those that moved as a single population.
}
\keyword{Population_Dynamics}
