% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_GenPopulation.R
\name{GetTerritory}
\alias{GetTerritory}
\title{Get Territory}
\usage{
GetTerritory(local, open)
}
\arguments{
\item{local}{Territories that are within reach of the target territory.}

\item{open}{Which territoies can be migrated to (i.e. no other population currently resides there).}
}
\description{
Returns a territory a population can migrate to or NA if none are available.
}
\keyword{Population_Dynamics}
