% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{GroupBySeed}
\alias{GroupBySeed}
\title{Group By Seed}
\usage{
GroupBySeed(P, Data)
}
\arguments{
\item{P}{A list of parameters.}

\item{Data}{The Pre or Post output from an HBC simulation.}
}
\description{
Returns the territories descended from each seed.  Works for all sims, but lacks ancestory data.
}
\keyword{Plotting}
