% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_ParametersSimulation.R
\name{HCBAlternatorSimmulation}
\alias{HCBAlternatorSimmulation}
\title{Human Cultural Boundaries Alternator Simulation}
\usage{
HCBAlternatorSimmulation(P)
}
\arguments{
\item{P}{A list of parameters.}
}
\description{
Runs a simulation where each time step alternates between migration and horizontal transfer.
}
\keyword{SimParam}
