% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Horizontal.R
\name{HoritontalTransferRepeater}
\alias{HoritontalTransferRepeater}
\title{Horitontal Transfer Repeater}
\usage{
HoritontalTransferRepeater(P, S)
}
\arguments{
\item{P}{A list of parameters.}

\item{S}{A list of the data structures.}
}
\description{
A wrapper for the horizontal transfer process.  After migration, allow populations to exchnage phoneme information, losing or gaining syllables based on other populations in the simulation.  Occurs HSims number of time steps.
}
\keyword{Horizontal}
