% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Horizontal.R
\name{HorizontalTransfer}
\alias{HorizontalTransfer}
\title{Horizontal Transfer}
\usage{
HorizontalTransfer(P, languages, local, phonemeRelatedness, phonemeProbab,
  index)
}
\arguments{
\item{P}{A list of parameters.}

\item{languages}{All languages.}

\item{local}{Locality.}

\item{phonemeRelatedness}{The phoneme relatedness list.}

\item{phonemeProbab}{The probability of gaining each phoneme in the population.}

\item{index}{The target territory whose language may change.}
}
\description{
A function wrapper that get the language to modify and allows the phoneme change to either add/shift or remove a phoneme if this can be done.
}
\keyword{Horizontal}
