% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Initialize.R
\name{Initialize}
\alias{Initialize}
\title{Initialize}
\usage{
Initialize(P)
}
\arguments{
\item{P}{A list of parameters.}
}
\description{
The function wrapper that makes calls to create the population and phoneme data structures and then populatas them with initial data.
}
\keyword{Wrapper}
