% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Horizontal.R
\name{Lose}
\alias{Lose}
\title{Lose Phoneme}
\usage{
Lose(P, language, phonemeProbab)
}
\arguments{
\item{P}{A list of parameters.}

\item{language}{The target language to be modified if possible.}

\item{phonemeProbab}{The probability of gaining each phoneme in the population.}
}
\description{
Allows a language to either lose a phoneme to better match other populations.
}
\keyword{Horizontal}
