% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{MakeDistanceMap}
\alias{MakeDistanceMap}
\title{Make Distance Map}
\usage{
MakeDistanceMap(P)
}
\arguments{
\item{P}{A list of parameters.}
}
\description{
Creates a distance map based on the euclidian diatcnes between territories.
}
\keyword{Plotting}
