% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_GenLanguage.R
\name{MakeLanguage}
\alias{MakeLanguage}
\title{Make Language}
\usage{
MakeLanguage(P, phonemeProbab, phonemeRelatedness, language, popSize)
}
\arguments{
\item{P}{A list of parameters.}

\item{phonemeProbab}{The probability of gaining each phoneme in the population.}

\item{phonemeRelatedness}{The phoneme relatedness list.}

\item{language}{The parent langauge to mutate into a new language.}

\item{popSize}{The number of individuals in the parent poultaion.}
}
\description{
Copies and mutates the parent's langauge to create a new language for a founder party.
}
\keyword{Language}
