% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Migration.R
\name{Migration}
\alias{Migration}
\title{Migration}
\usage{
Migration(P, S)
}
\arguments{
\item{P}{A list of parameters.}

\item{S}{A list of data structures.}
}
\description{
Main simulation function.  Allows populations to migrate, split, and die.
}
\keyword{SimParam}
