% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{MigrationPlot}
\alias{MigrationPlot}
\title{Migration Plot}
\usage{
MigrationPlot(P, Data, groups = NA, colorSet = NA)
}
\arguments{
\item{P}{A list of parameters.}

\item{Data}{The Pre or Post output from an HBC simulation.}
}
\description{
Shows the expansion of populations from the seed population.  Ony works when Uproot and Death are FALSE.
}
\keyword{Plotting}
