% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Directions.R
\name{NextStepDirections}
\alias{NextStepDirections}
\title{Next Step Directions
Expands the Steps list one more step out.}
\usage{
NextStepDirections(firstStep, currentStep, start = 0)
}
\arguments{
\item{firstStep}{The original StepOne.}

\item{currentStep}{StepOne in its current state.}

\item{start}{How much to offset numbers (for phoneme structures).}
}
\description{
Next Step Directions
Expands the Steps list one more step out.
}
\keyword{Directions}
