% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Directions.R
\name{OneStepDirections}
\alias{OneStepDirections}
\title{One Step Directions}
\usage{
OneStepDirections(R, C, start = 0, round = FALSE)
}
\arguments{
\item{R}{The number of rows.}

\item{C}{The number of columns.}

\item{start}{How much to offset numbers (for phoneme structures).}

\item{round}{whether to make the spacing Round (Phonemes) or Sqaure (Territories).}
}
\description{
One Step Directions
}
\keyword{Directions}
