% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{PhonemeFrequencyPlots}
\alias{PhonemeFrequencyPlots}
\title{Phoneme Frequency Plots}
\usage{
PhonemeFrequencyPlots(P, Data, groups = NA, colorSet = NA)
}
\arguments{
\item{P}{A list of parameters.}

\item{Data}{The Pre or Post output from an HBC simulation.}
}
\description{
Shows how common each phonemes is in the simulation color coded by population.
}
\keyword{Plotting}
