% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{PhonemeMantel}
\alias{PhonemeMantel}
\title{Phoneme Mantel}
\usage{
PhonemeMantel(P, Data, repeats = 100)
}
\arguments{
\item{P}{A list of parameters.}

\item{Data}{The Pre or Post output from an HBC simulation.}

\item{repeats}{How many times to repeat the analysis.}
}
\description{
Performs both a Hamming and Jaccard Mantel test.
}
\keyword{Plotting}
