% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{PhonemePopulationFrequencyPlots}
\alias{PhonemePopulationFrequencyPlots}
\title{Phoneme Population Frequency Plots}
\usage{
PhonemePopulationFrequencyPlots(P, Data, groups = NA, colorSet = NA,
  sort = TRUE)
}
\arguments{
\item{P}{A list of parameters.}

\item{Data}{The Pre or Post output from an HBC simulation.}

\item{groups}{Group structure of which territories were descended from what population seed.}

\item{colorSet}{The colors to use.}

\item{sort}{Whether to sort the data from most to least frequent phoneme.}
}
\description{
Shows how common each phonemes is in the simulation color coded by population.
}
\keyword{Plotting}
