% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{PopulationPlot}
\alias{PopulationPlot}
\title{Population Plot}
\usage{
PopulationPlot(P, seedGroups, colors)
}
\arguments{
\item{P}{A list of parameters.}

\item{seedGroups}{Group structure of which territories were descended from what population seed.}

\item{colors}{A vector of colors, one for each seed.}
}
\description{
Population Plot
}
\keyword{Plotting}
