% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Barriers.R
\name{RemoveVerticalConnections}
\alias{RemoveVerticalConnections}
\title{Remove Vertical Connections}
\usage{
RemoveVerticalConnections(R, index, firstStep, above = TRUE,
  below = TRUE)
}
\arguments{
\item{R}{The number of rows in the population matrix.}

\item{index}{The target territory.}

\item{firstStep}{The local directions created by OneStepDirections().}

\item{above}{Whether to remove the upper diagonal.}

\item{below}{Whether to remove the lower diagonal.}
}
\description{
Affects local territories right/East (and perhaps to the Northeast and Southeast) the target territory (index) and left/West (perhaps Northwest and Southwest) of index + R.
}
\keyword{Barriers}
