% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_GenPopulation.R
\name{ResetPopulation}
\alias{ResetPopulation}
\title{Reset Population}
\usage{
ResetPopulation()
}
\description{
Resets phoeneme and population data in the original territory when an entire population moes to a new territory.
}
\keyword{Population_Dynamics}
