% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Directions.R
\name{ShiftDirections}
\alias{ShiftDirections}
\title{Shift Directions}
\usage{
ShiftDirections(nPhonemes, start = 0)
}
\arguments{
\item{nPhonemes}{The number of phonemes.}

\item{start}{Where to start number the phonemes (0 for consonants, number of consonants +1 for vowels).}
}
\description{
Returns the relationships between phonemes with an offset of start.
}
\keyword{Directions}
