% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{SnapshotPlot}
\alias{SnapshotPlot}
\title{Snapshot Plot}
\usage{
SnapshotPlot(P, Data, colors)
}
\arguments{
\item{P}{A list of parameters.}

\item{Data}{The Pre or Post output from an HBC simulation.}

\item{colors}{A vector of colors, one for each seed.}
}
\description{
Shows which territories are populated and from which seed they decended.
}
\keyword{Plotting}
