% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_GenPopulation.R
\name{UpdateStructuresRemove}
\alias{UpdateStructuresRemove}
\title{Update Structures Remove}
\usage{
UpdateStructuresRemove(S, remove)
}
\arguments{
\item{S}{A list of data structures.}

\item{Remove}{The indicies of territories whose data should be erased.}
}
\description{
Deletes population and language information for specified territories.
}
\keyword{Population_Dynamics}
