% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_ParametersSimulation.R
\name{HCBSimmulation}
\alias{HCBSimmulation}
\title{Human Cultural Boundaries Simulation}
\usage{
HCBSimmulation(P)
}
\arguments{
\item{P}{A list of parameters.}
}
\description{
Runs a simulation where migration and horizontal transfer are completely separate.
}
\keyword{SimParam}
